pretty: Allow piping the next placeholder to a command
Add a new placeholder for pretty formats "%|" that takes a command as an
argument and then pipes the result of the next placeholder through it.

Imagine a use case such as the following.  A git repo exists with some
custom notes on the "ci/status" ref.  If a note exists for any given
commit it will be a 1 if the commit passed the test suite or 0 if the
commit failed.

A script called "translate" will filter the 1, 0, or empty stream and
convert it to a "pass" or "fail" or four empty spaces....and then can be
used such as:

$ git log --notes=ci/status --decorate=auto --pretty='%h [ci: %|(./translate)%N]%d'
5adfd1a [ci:     ] (HEAD -> master)
adf318b [ci: pass] (origin/master, temp)
5ad36ad [ci: pass]
8c6ca38 [ci: pass] (local/master)
7c95c43 [ci: fail]
ec1688e [ci:     ]
3073e65 [ci:     ]
7e64add [ci:     ]
5271413 [ci:     ]

Signed-off-by: Jacob Gamble <gamblejr@prototypeind.com>